Refactor: remove dayjs and replace with luxon
