+++ src/components/PageComponents/ScheduleComponents/Sidebar/SmallCalendar/DayPicker/index.js
@@ -1,0 +1,3 @@
+import DayPicker from './DayPicker';
+
+export default DayPicker;
