+++ src/components/PageComponents/ScheduleComponents/Sidebar/SmallCalendar/SmallCalendar.jsx
@@ -3,65 +3,7 @@ import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { AdapterLuxon } from '@mui/x-date-pickers/AdapterLuxon';
 import { DateCalendar } from '@mui/x-date-pickers';
 import PropTypes from 'prop-types';
-import { styled } from '@mui/material/styles';
-import { PickersDay } from '@mui/x-date-pickers/PickersDay';
-import dayjs from 'dayjs';
-import isBetweenPlugin from 'dayjs/plugin/isBetween';
-import isoWeek from 'dayjs/plugin/isoWeek';
-
-dayjs.extend(isBetweenPlugin);
-dayjs.extend(isoWeek);
-
-const CustomPickersDay = styled(PickersDay, {
-  shouldForwardProp: (prop) => prop !== 'isSelected' && prop !== 'isHovered',
-})(({ theme, isSelected, isHovered }) => ({
-  borderRadius: '50%',
-  margin: '0 2px',
-  ...(isSelected && {
-    backgroundColor: theme.palette.primary.main,
-    color: theme.palette.primary.contrastText,
-    '&:hover, &:focus': {
-      backgroundColor: theme.palette.primary.main,
-    },
-  }),
-  ...(isHovered && {
-    backgroundColor: theme.palette.primary.light,
-    '&:hover, &:focus': {
-      backgroundColor: theme.palette.primary.light,
-    },
-    ...theme.applyStyles('dark', {
-      backgroundColor: theme.palette.primary.dark,
-      '&:hover, &:focus': {
-        backgroundColor: theme.palette.primary.dark,
-      },
-    }),
-  }),
-}));
-
-const isInSameWeek = (dayA, dayB) => {
-  if (dayB == null) {
-    return false;
-  }
-  const dayADayjs = dayjs(dayA.toJSDate());
-  const dayBDayjs = dayjs(dayB.toJSDate());
-  return dayADayjs.isoWeek() === dayBDayjs.isoWeek() && dayADayjs.isoWeekday() !== 8;
-};
-
-function Day(props) {
-  // eslint-disable-next-line react/prop-types
-  const { day, selectedDay, hoveredDay, ...other } = props;
-
-  return (
-    <CustomPickersDay
-      {...other}
-      day={day}
-      disableMargin
-      selected={false}
-      isSelected={isInSameWeek(day, selectedDay)}
-      isHovered={isInSameWeek(day, hoveredDay)}
-    />
-  );
-}
+import DayPicker from './DayPicker';
 
 const SmallCalendar = ({ selectedDate, handleDateChange }) => {
   const [hoveredDay, setHoveredDay] = useState(null);
@@ -72,7 +14,7 @@ const SmallCalendar = ({ selectedDate, handleDateChange }) => {
         showDaysOutsideCurrentMonth
         value={selectedDate}
         onChange={handleDateChange}
-        slots={{ day: Day }}
+        slots={{ day: DayPicker }}
         slotProps={{
           day: (ownerState) => ({
             selectedDay: selectedDate,
@@ -85,9 +27,10 @@ const SmallCalendar = ({ selectedDate, handleDateChange }) => {
     </LocalizationProvider>
   );
 };
-export default SmallCalendar;
 
 SmallCalendar.propTypes = {
   selectedDate: PropTypes.object,
   handleDateChange: PropTypes.func,
 };
+
+export default SmallCalendar;
